requirements.txt
Flask==0.12.3
Werkzeug==0.12.1
gunicorn==19.7.1
Jinja2==2.9.6
arrow==0.10.0
psycopg2-binary==2.7.7
SQLAlchemy==1.1.9
Flask-Login==0.4.0
pytest==4.3.0